inseridas as anotações nos DTO's Item EntregaRequest e no EntregaRequest associado o cliente a entrega

+++ src/main/java/com/logistica/entregas_api/repository/EntregaRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EntregaRepository extends JpaRepository<Entrega, Integer> {
+public interface EntregaRepository extends JpaRepository<Entrega, Long> {
 
-    List<Entrega> findByClienteId(Integer clienteId);
+    List<Entrega> findByClienteId(Long clienteId);
 
 }
